requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
psycopg==3.1.16
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
psycopg==3.1.16
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
